requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==3.2.16
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.3.0
importlib-metadata==6.0.0
mongoengine==0.26.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pymongo==4.3.3
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.4.0
zipp==3.12.0
